KAN-39: Reorganise file structure
