--- frontend/noinfluence/src/authentication/CognitoConfig.js
@@ -1,7 +1,0 @@
-export const cognitoConfig = {
-  domain: process.env.DOMAIN,
-  clientId: process.env.CLIENT_ID,
-  redirectUri: process.env.REDIRECT_URI,
-  region: process.env.REGION,
-  scope: process.env.SCOPE,
-};
