+++ frontend/src/components/authentication/CognitoConfig.js
@@ -1,0 +1,7 @@
+export const cognitoConfig = {
+  domain: process.env.REACT_APP_DOMAIN,
+  clientId: process.env.REACT_APP_CLIENT_ID,
+  redirectUri: process.env.REACT_APP_REDIRECT_URI,
+  region: process.env.REACT_APP_REGION,
+  scope: process.env.REACT_APP_SCOPE,
+};
