+++ frontend/src/components/authentication/ProtectedRoute.js
