+++ frontend/src/components/index.js
@@ -1,3 +1,4 @@
+export * from './authentication';
 export * from './HomeDashBoard';
 export * from "./NavigationBar";
 export * from './SpaceDashBoard';
